requirements.txt
gradio
langchain-predictionguard
IPython
umap-learn
pytubefix
youtube_transcript_api
torch
transformers
matplotlib
seaborn
datasets